# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Boardly API",
    "description": "칸반 보드 스타일의 작업 관리 웹 애플리케이션 API",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "security": [
    {
      "oauth2": [
        "read",
        "write",
        "openid"
      ]
    }
  ],
...
  "paths": {
...
    "/api/boards/{boardId}": {
      "get": {
        "tags": [
          "Board"
        ],
        "summary": "보드 상세 조회",
        "description": "보드의 상세 정보를 조회합니다. 보드의 컬럼, 카드, 멤버, 라벨 등의 모든 정보를 포함합니다.",
        "operationId": "getBoardDetail",
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "조회할 보드 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "보드 상세 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDetailResponse"
                }
              }
            }
          },
          "403": {
            "description": "보드 접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "보드를 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2": [
              "read",
              "openid"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "Board"
        ],
        "summary": "보드 업데이트",
        "description": "기존 보드의 제목과 설명을 업데이트합니다.",
        "operationId": "updateBoard",
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "업데이트할 보드 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "보드 업데이트 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 데이터",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "보드 수정 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "보드를 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "아카이브된 보드는 수정 불가",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 값이 유효하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2": [
              "write",
              "openid"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Board"
        ],
        "summary": "보드 삭제",
        "description": "보드를 영구적으로 삭제합니다. 보드와 관련된 모든 데이터(리스트, 카드, 멤버)가 함께 삭제됩니다.",
        "operationId": "deleteBoard",
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "삭제할 보드 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "보드 삭제 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 데이터",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "보드 삭제 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "보드를 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 값이 유효하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2": [
              "write",
              "openid"
            ]
          }
        ]
      }
    },
    "/api/board-lists/{listId}": {
      "put": {
        "tags": [
          "BoardList"
        ],
        "summary": "보드 리스트 수정 (Update)",
        "description": "특정 보드 리스트의 정보를 수정합니다.",
        "operationId": "updateBoardList",
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "description": "수정할 리스트 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoardListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "보드 리스트 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardListResponse"
                }
              }
            }
          },
          "403": {
            "description": "보드 접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "보드 또는 리스트를 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 값이 유효하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2": [
              "read",
              "write",
              "openid"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "BoardList"
        ],
        "summary": "보드 리스트 삭제 (Delete)",
        "description": "특정 보드 리스트를 삭제합니다.",
        "operationId": "deleteBoardList",
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "description": "삭제할 리스트 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "보드 리스트 삭제 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "description": "보드 접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "보드 또는 리스트를 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 값이 유효하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2": [
              "write",
              "openid"
            ]
          }
        ]
      }
    },
    "/api/board-lists/{listId}/position": {
      "put": {
        "tags": [
          "BoardList"
        ],
        "summary": "보드 리스트 위치 변경 (Update Position)",
        "description": "특정 보드 리스트의 위치를 변경합니다.",
        "operationId": "updateBoardListPosition",
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "description": "위치를 변경할 리스트 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoardListPositionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "보드 리스트 위치 변경 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardListResponse"
                }
              }
            }
          },
          "403": {
            "description": "보드 접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "보드 또는 리스트를 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 값이 유효하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2": [
              "write",
              "openid"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "path": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            }
          },
          "context": {}
        }
      },
...
      "UpdateBoardRequest": {
        "type": "object",
        "description": "보드 업데이트 요청 정보",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "BoardResponse": {
        "type": "object",
        "properties": {
          "boardId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isArchived": {
            "type": "boolean"
          },
          "ownerId": {
            "type": "string"
          },
          "isStarred": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "BoardDetailResponse": {
        "type": "object",
        "properties": {
          "boardId": {
            "type": "string"
          },
          "boardName": {
            "type": "string"
          },
          "boardDescription": {
            "type": "string"
          },
          "isStarred": {
            "type": "boolean"
          },
          "boardColor": {
            "type": "string"
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardColumnResponse"
            }
          },
          "boardMembers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardMemberResponse"
            }
          },
          "labels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardLabelResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://localhost:8080/oauth2/authorize",
            "tokenUrl": "http://localhost:8080/oauth2/token",
            "scopes": {
              "read": "Read access",
              "write": "Write access",
              "openid": "OpenID Connect"
            }
          }
        }
      }
    }
  }
}